Create Q2
Create a class by name QuizMarks with the instance variables – StudentName, QuestionCount that stores the total number of questions, CorrectCount that stores the number of question answered correctly, WrongCount and TotalMarks. Include a default and a parameterized constructor to initialize StudentName, QuestionCount and CorrectCount.

Create an interface by name DetermineTotal with a method calcTotal().

Create a class by name Result. The class should extend QuizMarks and also implement the DetermineTotal interface. Include a default and a parameterized constructor to initialize the instance members of the base class. The constructors should have an explicit call to the base class constructor. Also override the calcTotal() method which first calculates the WrongCount as the difference between QuestionCount and CorrectCount and then calculate the total marks as follows. Award 2 marks for every correct answer and reduce 1 marks for every wrong answer.

Write a main class that tests the above hierarchy for an array of 'n' objects.